# old/api/docs/swagger.yaml
definitions:
...
  github_com_nicolailuther_butter_internal_dto.ClientInsightsResponse:
    properties:
      convertion_rate:
        type: number
      cpm:
        type: number
      gross_revenue:
        type: number
      media_views:
        type: integer
      net_revenue:
        type: number
      onlyfans_accounts_count:
        type: integer
      onlyfans_views:
        type: integer
      tracking_links_clicks:
        type: integer
      tracking_links_subscribers:
        type: integer
    type: object
  github_com_nicolailuther_butter_internal_dto.ClientPostingInsights:
    properties:
      accounts_by_platform:
        additionalProperties:
          type: integer
        type: object
      daily_goal_total:
        type: integer
      daily_posts:
        type: integer
      weekly_goal_total:
        type: integer
      weekly_posts:
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_enums.ClientIndustry:
    enum:
    - only fans
    - credit stacking
    - real estate
    - ""
    type: string
    x-enum-varnames:
    - ClientIndustryOnlyFans
    - ClientIndustryCreditStacking
    - ClientIndustryRealEstate
    - ClientUnknownIndustry
...
  github_com_nicolailuther_butter_internal_models.Client:
    properties:
      company_percentage:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      industry:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_enums.ClientIndustry'
      name:
        type: string
      profile_image:
        $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.File'
      profile_image_id:
        type: integer
      updated_at:
        type: string
      user_id:
        description: Should be not null
        type: integer
    type: object
...
  github_com_nicolailuther_butter_internal_models.File:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      etag:
        type: string
      id:
        type: integer
      path:
        type: string
      size:
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  github_com_nicolailuther_butter_internal_responses.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    name: Default
    url: https://default.dev
  description: Backend service for Butter
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Butter API
  version: "1.0"
paths:
...
  /api/v1/clients:
    get:
      description: Retrieve all clients assined to a specific user
      parameters:
      - description: Get by user ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of assigned clients
          schema:
            items:
              $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Assigned Clients by User
      tags:
      - clients
    post:
      consumes:
      - multipart/form-data
      description: Create a new Client with optional profile picture file upload
      parameters:
      - description: Profile image file
        in: formData
        name: profile_image
        type: file
      - in: formData
        name: company_percentage
        type: integer
      - enum:
        - only fans
        - credit stacking
        - real estate
        - ""
        in: formData
        name: industry
        type: string
        x-enum-varnames:
        - ClientIndustryOnlyFans
        - ClientIndustryCreditStacking
        - ClientIndustryRealEstate
        - ClientUnknownIndustry
      - in: formData
        name: name
        type: string
      - in: formData
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client created successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Client
      tags:
      - clients
  /api/v1/clients/{id}:
    delete:
      description: Delete a client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Client
      tags:
      - clients
    get:
      description: Retrieve a client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Client by ID
      tags:
      - clients
    patch:
      consumes:
      - multipart/form-data
      description: Update a client by ID with optional profile picture file upload
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile image file
        in: formData
        name: profile_image
        type: file
      - in: formData
        name: company_percentage
        type: integer
      - enum:
        - only fans
        - credit stacking
        - real estate
        - ""
        in: formData
        name: industry
        type: string
        x-enum-varnames:
        - ClientIndustryOnlyFans
        - ClientIndustryCreditStacking
        - ClientIndustryRealEstate
        - ClientUnknownIndustry
      - in: formData
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client updated successfully
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Client
      tags:
      - clients
  /api/v1/clients/{id}/insights:
    get:
      description: Retrieve insights for a specific client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date for insights (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date for insights (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client insights data
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.ClientInsightsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Client Insights
      tags:
      - clients
  /api/v1/clients/{id}/posting-insights:
    get:
      description: Retrieve posting insights for a specific client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Client posting insights data
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_dto.ClientPostingInsights'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_nicolailuther_butter_internal_responses.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Client Posting Insights
      tags:
      - clients
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  PushApiKey:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"